# decant-standalone/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Decant API
  description: |
    RESTful API for managing hierarchical knowledge graphs with AI-powered classification and enrichment.

    Decant provides dual-hierarchy organization (functional and organizational), automatic content
    classification using GPT-4o-mini, and real-time updates via Server-Sent Events.

    ## Key Features
    - Dual hierarchy organization (function & organization)
    - AI-powered content classification
    - Phase 2 enrichment for deep analysis
    - Full-text search with advanced filters
    - Real-time notifications via SSE
    - Backup and restore capabilities
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: http://localhost:{port}
    description: Configurable port
    variables:
      port:
        default: '3000'
        description: Server port
...
paths:
...
  /api/import:
    post:
      tags: [Import]
      summary: Import URL
      description: Import a URL with automatic AI classification
      operationId: importUrl
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportUrlInput'
      responses:
        '200':
          description: Import successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '408':
          $ref: '#/components/responses/RequestTimeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/import/check:
    get:
      tags: [Import]
      summary: Check import status
      description: Check if a URL has already been imported
      operationId: checkImportStatus
      parameters:
        - name: url
          in: query
          required: true
          description: URL to check
          schema:
            type: string
            format: uri
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportCheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================
  # Queue API
  # ============================================================
...
  /api/settings/api-key:
    post:
      tags: [Settings]
      summary: Set API key
      description: Configure the OpenAI API key
      operationId: setApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apiKey]
              properties:
                apiKey:
                  type: string
                  minLength: 20
                  example: sk-proj-abc123...
      responses:
        '200':
          description: API key configured successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Settings]
      summary: Delete API key
      description: Remove the stored OpenAI API key
      operationId: deleteApiKey
      responses:
        '200':
          description: API key deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/settings/api-key/status:
    get:
      tags: [Settings]
      summary: Get API key status
      description: Check if an OpenAI API key is configured
      operationId: getApiKeyStatus
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  configured:
                    type: boolean
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================
  # Events API
  # ============================================================
...
components:
...
  schemas:
...
    Node:
      type: object
      required:
        - id
        - title
        - url
        - source_domain
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        url:
          type: string
          format: uri
        source_domain:
          type: string
        company:
          type: string
          nullable: true
        phrase_description:
          type: string
          nullable: true
        short_description:
          type: string
          nullable: true
        logo_url:
          type: string
          format: uri
          nullable: true
        ai_summary:
          type: string
          nullable: true
        extracted_fields:
          type: object
          additionalProperties: true
        metadata_tags:
          type: array
          items:
            type: string
        key_concepts:
          type: array
          items:
            type: string
        function_parent_id:
          type: string
          format: uuid
          nullable: true
        organization_parent_id:
          type: string
          format: uuid
          nullable: true
        function_code:
          type: string
          nullable: true
        organization_code:
          type: string
          nullable: true
        phase2_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ImportUrlInput:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
        forceRefresh:
          type: boolean
          default: false
        priority:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: string
...
    ImportResponse:
      type: object
      properties:
        success:
          type: boolean
        nodeId:
          type: string
          format: uuid
        cached:
          type: boolean
        node:
          $ref: '#/components/schemas/Node'
        classification:
          type: object
          properties:
            segment:
              type: string
            category:
              type: string
            contentType:
              type: string
            organization:
              type: string
            confidence:
              type: number
              format: float
        hierarchyCodes:
          type: object
          properties:
            function:
              type: string
            organization:
              type: string
        metadata:
          type: object
        phase2:
          type: object
          properties:
            queued:
              type: boolean
            jobId:
              type: string
              format: uuid
    ImportCheckResponse:
      type: object
      properties:
        exists:
          type: boolean
        cached:
          type: boolean
        nodeId:
          type: string
          format: uuid
        classification:
          type: object
        cachedAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RequestTimeout:
      description: Request Timeout
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PayloadTooLarge:
      description: Payload Too Large
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Too Many Requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadGateway:
      description: Bad Gateway
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Service Unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication (future enhancement)